requirements.txt
opencv_python==4.4.0.44
telepot==12.7
pillow==8.0.1